A working simulation of branching branches
This point in the project is a working simulation of a "plant". A stem will grow with 2-8 branches that each keep branching with the same number of branches.
Still to do:
Limit branch angles.
Add an way to interact: removing and adjusting branches.
Leaves. Sun?
Varying branch sizes.
Energy? This would come from light and leaves and be expended for branch size and length.